# OMT packing for rtree.