--- a/features/car/destination.feature
+++ b/features/car/destination.feature
@@ -8,7 +8,7 @@ Feature: Car - Destination only, no passing through
         Given the node map
             """
             a       e
-              b c d
+              b2c1d
 
             x       y
             """
@@ -24,18 +24,18 @@ Feature: Car - Destination only, no passing through
             | from | to | route      |
             | a    | b  | ab,ab      |
             | a    | c  | ab,bcd     |
-            | a    | d  | ab,bcd,bcd |
+            | a    | 1  | ab,bcd,bcd |
             | a    | e  | axye,axye  |
             | e    | d  | de,de      |
             | e    | c  | de,bcd     |
-            | e    | b  | de,bcd,bcd |
+            | e    | 2  | de,bcd,bcd |
             | e    | a  | axye,axye  |
 
     Scenario: Car - Destination only street
         Given the node map
             """
             a       e
-              b c d
+              b2c1d
 
             x       y
             """
@@ -52,11 +52,11 @@ Feature: Car - Destination only, no passing through
             | from | to | route       |
             | a    | b  | ab,ab       |
             | a    | c  | ab,bc       |
-            | a    | d  | ab,cd       |
+            | a    | 1  | ab,cd       |
             | a    | e  | axye,axye   |
             | e    | d  | de,de       |
             | e    | c  | de,cd       |
-            | e    | b  | de,bc       |
+            | e    | 2  | de,bc       |
             | e    | a  | axye,axye   |
 
     Scenario: Car - Routing inside a destination only area
@@ -117,7 +117,7 @@ Feature: Car - Destination only, no passing through
                    +    \
                    +    |
                    d    |
-                    \___e
+                    1___e
             """
 
         And the ways
@@ -129,5 +129,5 @@ Feature: Car - Destination only, no passing through
         When I route I should get
             | from | to | route         |
             | e    | a  | acbe,acbe     |
-            | d    | a  | de,acbe,acbe  |
+            | 1    | a  | de,acbe,acbe  |
             | c    | d  | cd,cd         |