--- a/src/extractor/extractor.cpp
+++ b/src/extractor/extractor.cpp
@@ -564,7 +564,7 @@ void Extractor::BuildRTree(std::vector<EdgeBasedNode> node_based_edge_list,
     node_based_edge_list.resize(new_size);
 
     TIMER_START(construction);
-    util::StaticRTree<EdgeBasedNode> rtree(node_based_edge_list,
+    util::StaticRTree<EdgeBasedNode> rtree(std::move(node_based_edge_list),
                                            config.rtree_nodes_output_path,
                                            config.rtree_leafs_output_path,
                                            coordinates);