… src/benchmarks/route.cpp
… src/benchmarks/table.cpp
… src/tools/rtree-dump.cpp