# Issue

This PR is to address https://github.com/Project-OSRM/osrm-backend/issues/3378.  Our current RTree uses a Hilbert sort to group leaf nodes.  For road network data, this gives us a less-than-optimal packing with quite a lot of overlap for leaf nodes.  We can speed up coordinate snapping by using a better packing method.

This PR implements the [Overlap Minimizing Top-down bulk-loading](http://ftp.informatik.rwth-aachen.de/Publications/CEUR-WS/Vol-74/files/FORUM_18.pdf) in our rtree, which produces an RTree with far less overlap than the current Hilbert Sort approach.

As a bonus, OMT makes it easy to back the TreeNode data into a breadth-first-search linear order, which means we don't need fixed-sized child arrays in the TreeNode struct, range pointers are sufficient.  This greatly reduces the size of the `.ramIndex` files.

## Tasklist
 - [ ] compare to Boost's packing approach, which is an [altered STR method](https://github.com/boostorg/geometry/blob/develop/include/boost/geometry/index/detail/rtree/pack_create.hpp#L110)
 - [ ] benchmarking
 - [ ] add regression / cucumber cases (see docs/testing.md)
 - [ ] review
 - [ ] adjust for comments