@danpat 2017-04-14
Just to confirm that this avenue is worth pursuing, I did some quick profiling on the existing RTree using `src/benchmarks/static_rtree.cpp`.

This code runs two test scenarios - each takes about 3.5 seconds on my machine, single threaded.

Below shows the breakdown of one of the scenarios.  The key observation is that we're spending about 2.84 seconds out of 3.49 seconds (81%) of our time in the `ExploreLeafNode` method.  Leaf-node exploration is exactly what STR should reduce.

<img width="1007" alt="screen shot 2017-04-13 at 10 35 51 pm" src="https://cloud.githubusercontent.com/assets/1892250/25034273/ef1aef9a-2099-11e7-9856-1d404fe6d6ed.png">