@danpat 2017-04-14
Also benchmarked the table plugin for a `/table` request with 1000 nodes:

<img width="992" alt="screen shot 2017-04-13 at 11 33 55 pm" src="https://cloud.githubusercontent.com/assets/1892250/25035323/bd390ffe-20a1-11e7-9650-3ed545997bf5.png">

At this size, we're spending 13% of our time in the RTree search.

For a 100 coordinate `/table`, it's 15%:

<img width="998" alt="screen shot 2017-04-13 at 11 36 09 pm" src="https://cloud.githubusercontent.com/assets/1892250/25035359/05734e74-20a2-11e7-9bb9-1d2975741ee2.png">

And for 500 coordinates:

<img width="1012" alt="screen shot 2017-04-13 at 11 37 39 pm" src="https://cloud.githubusercontent.com/assets/1892250/25035390/3b0377b2-20a2-11e7-9925-5b4bc77e052d.png">

(Note for the curious: these results are from the Timing component of `Instruments.app` on OSX (part of XCode))

/cc @daniel-j-h