@danpat 2017-04-25
Hokay, assuming I haven't messed this up:

Before:

```
$ ls -l san-francisco_california.osrm.*Index
-rw-r--r--  1 danpat  staff  14045184 Apr 22 13:07 san-francisco_california.osrm.fileIndex
-rw-r--r--  1 danpat  staff     14912 Apr 22 13:07 san-francisco_california.osrm.ramIndex
$ ./src/benchmarks/rtree-bench san-francisco_california.osrm.ramIndex san-francisco_california.osrm.fileIndex san-francisco_california.osrm.nodes
Running raw RTree queries (1 result) with 10000 coordinates: Took 2.67895 seconds (2678.95ms)  ->  0.267895 ms/query (2678.95ms)
Running raw RTree queries (10 results) with 10000 coordinates: Took 3.53938 seconds (3539.38ms)  ->  0.353938 ms/query (3539.38ms)
```

After:

```
$ ls -l san-francisco_california.osrm.*Index
-rw-r--r--  1 danpat  staff  33554432 Apr 25 01:20 san-francisco_california.osrm.fileIndex
-rw-r--r--  1 danpat  staff      3160 Apr 25 01:20 san-francisco_california.osrm.ramIndex
$ ./src/benchmarks/rtree-bench san-francisco_california.osrm.ramIndex san-francisco_california.osrm.fileIndex san-francisco_california.osrm.nodes
Running raw RTree queries (1 result) with 10000 coordinates: Took 1.03206 seconds (1032.06ms)  ->  0.103206 ms/query (1032.06ms)
Running raw RTree queries (10 results) with 10000 coordinates: Took 1.83538 seconds (1835.38ms)  ->  0.183538 ms/query (1835.38ms)
```

The `.ramIndex` file is only 21% of the previous size - this is a legit space saving and means we can use a much smaller `BRANCHING_FACTOR` (and probably where much of the performance increase comes from, but not all).

Speed wise, it looks like queries only take 38% of the time they previously did.  Not a bad speedup at all.

I'm not quite sure why the `.fileIndex` has doubled - I suspect I've mis-calculated the `LEAF_NODE_PAGE_SIZE` and there is a lot of dead space.  Should be fixable.