@oxidase 2017-04-25
i have spotted in OMT leafs
![omt](https://cloud.githubusercontent.com/assets/4421046/25384705/9720801c-29c1-11e7-9ca0-aed3c1167728.png) that one leaf has large size and contains one distant node 
![omt1](https://cloud.githubusercontent.com/assets/4421046/25384745/b5bd76ba-29c1-11e7-8ff5-0df6ef4c5bbb.png)
Probably there is a problem with ceiling?

I have pushed the script that dumps r-trees at https://github.com/oxidase/home/blob/master/notes/topics/osrm/dump_rtree.py