@TheMarex 2017-04-25
Given the problem with leaf node packing maybe we can get around the concept of a `LeafNode` completely and only store a continuous array of `EdgeDataT` on disk:

1. Instead of saving the bounding box of each `TreeNode` we save the bounding boxes of each child: `std::array<Rectangle, BRANCHING_FACTOR> child_minimum_bounding_rectangle`
2. `first_child_index` already has the dual use of either pointing to a `LeafNode` or another `TreeNode`. Instead we could also make it point at an entry of `vector_view<EdgeDataT>`
3. The query doesn't need to change much, because we already need to fetch every childs MBR here https://github.com/Project-OSRM/osrm-backend/blob/master/include/util/static_rtree.hpp#L577-L583

This should also cache-optimize the query since processing a `TreeNode` is now memory local (child bounding boxes and IDs are already stored in the node).