@daniel-j-h 2017-05-17
@danpat can we get a small rundown of what's the problem here or why your approach didn't work out? I'd love to get some knowledge out of this pull request wrt. rtree mechanics.