@danpat 2017-05-17
@daniel-j-h the downside to OMT was the overall data size.  There was a performance gain, but it came at the expense of a fairly large increase in memory usage.  Because OMT uses a top-down packing approach, it was difficult to tweak the branching factor to get the right performance/size balance.  The OMT `.ramIndex` file was turning out 10x or larger.

Our current approach where the lowest level has many more branches than the majority of the tree seems to perform better overall, and requires less space (more branches at the leaf nodes = fewer nodes in the tree altogether = smaller trees).

The OMT construction method is also slow - about 5x slower than the Hilbert Sort.

I just opened up #4056 that ports some of the data structure changes from this PR into our existing Hilbert Sort structure - it speeds things up by >30% for a reasonable increase in RAM usage.

At this stage, I don't think OMT actually buys us anything.

Closing in favour of #4056